server_backend/Docker_not_tested/requirements_server.txt
# Установка зависимостей: pip install -r requirements_server.txt

flask
psycopg2-binary
